backend/requirements.txt
fastapi==0.104.1
starlette==0.27.0
uvicorn==0.24.0
click==8.1.7
pydantic==2.4.2
pydantic-core==2.10.1
typing-extensions==4.8.0
annotated-types==0.6.0
python-dotenv==1.0.0
sqlalchemy==2.0.9
pymongo==4.3.3
motor==3.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
anyio==3.7.1
sniffio==1.3.0
aiohttp==3.8.4
openai==1.11.0
google-generativeai==0.2.0
slowapi==0.1.8
limits==3.3.0
deprecated==1.2.14
wrapt==1.15.0
passlib[bcrypt]==1.7.4
importlib-resources==6.1.1
packaging==23.2
aiofiles==23.2.1
PyPDF2==3.0.1
pillow==10.1.0
transformers==4.50.0
requests==2.31.0
urllib3==2.0.7
torch==2.0.1
sympy==1.12
accelerate==0.23.0
